fix: Update README and add report input keys preprocessor
This PR updates the README with the instructions to build a local Docker images for testing purposes of the action. Also, it adds a filter for the REPORT_INPUT_KEYS env var that will act as a cleanser of the input data in case of a user error. Apart from that, it adds a uuid for the generated JSON payloads for the found vulnerabilities. Finally, a new input variable is introduced called RUNS_ON_GITHUB that will control the environment on which the action is used in order to make local testing on Docker containers easier.