feat: dirty-json addition
This PR adds the ability to cleanse and beautify the input JSON used for
the reporting tools trying to be integrated with Jira. Also, it helps
the user of this action a bit more since s/he doens't have to preprocess
the data to be imported in an extensive way.

Contributes to: camelotls/actions-jira-integration#67

Signed-off-by: Stelios Gkiokas <stelios.giokas@camelotls.com>