fix: Minimise the loss of issues
This PR attempts to minimise the loss of issues that are not raised due
to the inability to parse the providing JSON input despite the fact that
`dirty-json` is used. It has been noticed that the main problem lies to
cases in which boolean values are supplied in the JSON input, hence the
focus on this PR was on there.

Contributes to: camelotls/actions-jira-integration#93

Signed-off-by: Stelios Gkiokas <stelios.giokas@camelotls.com>