feat: Integration tests addition
This PR contains the integration tests for the Jira action which will
deploy a vanilla instance of Jira in a Docker container and attempt to
use the action to create and search for issues. This will be very
helpful in order to spot instantly when a PR breaks the intended
functionality of the action.

Contributes to: cameltols/actions-jira-integration#175

Signed-off-by: Konstantinos Perperidis <k.a.perperidis@gmail.com>
Signed-off-by: Stelios Gkiokas <s_giokas@hotmail.com>