feat: Flexible Template Creation (#285)
This PR provides a new approach on the way the input fields are passed from the user. We can now have a much more flexible way to construct the Jira issues payloads and leave to the user to choose some extra fields that we haven't initially thought of (e.g. the Epic Link, the Components etc).

Contributes to: camelotls/actions-jira-integration#257

Signed-off-by: George Papadakis <george.papadakis@camelotls.com>
Signed-off-by: Stelios Gkiokas <s_giokas@hotmail.com>