[NO JIRA] Add nullish operator in pair matcher
If any of the reportValue doesn't contain a {{field}} .match will
return null which causes the spread operator to throw an error
"TypeError: reportValue.match is not a function or its return value is
not iterable". Add a nullish coalescing operator and fallback to empty
Array if match return null.